+++ src/UnionArgParser/assemblyInfo.fs
@@ -1,9 +1,9 @@
 ﻿namespace System
 open System.Reflection
 
-[<assembly: AssemblyVersionAttribute("0.7.2")>]
-[<assembly: AssemblyFileVersionAttribute("0.7.2")>]
+[<assembly: AssemblyVersionAttribute("0.8.0")>]
+[<assembly: AssemblyFileVersionAttribute("0.8.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "0.7.2"
+    let [<Literal>] Version = "0.8.0"